requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
requests>=2.31.0
citeproc-py>=0.6.0
pydantic>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
httpx>=0.24.1
python-multipart>=0.0.6
pytest>=7.4.1
